Initial commit: Implement basic functionality and layout
- Added 'index.html', 'style.css', and 'script.js' files.
- Created 'images' folder to contain 'loader.gif' for the loader (work in progress).
- Added a name input and submit button to send the name to the APIs.
- Created the 'resultSquares' function to display the results of the APIs request.
- Implemented the 'insertButtons' function to create the label "Am I right?", the buttons "yes" and "no", and their respective messages.
- Improved styling needed (work in progress).

Overall, this commit adds the basic functionality and layout for the project.
The 'index.html' file contains the main structure of the page, 'style.css' contains the basic styling, and 'script.js' contains the JavaScript code necessary for the user interactions.
Additional work is needed to improve the styling, which is still in progress.
The 'images' folder contains the 'loader.gif' image used for the loader.
The 'name' input and 'resultSquares' function allow the user to interact with the APIs and see the results.
The 'insertButtons' function adds the "Am I right?" label and the "yes" and "no" buttons, along with their respective messages.